Installer/darwin,linux: Check VBOX_WITH_FE_BALLOONCTRL before assuming VBoxBalloonCtrl and service files are there. bugref:10030
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@98679 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Installer/linux/install.sh
@@ -332,7 +332,9 @@ if [ "$ACTION" = "install" ]; then
     ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/VBoxSDL
     ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/VBoxVRDP
     ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/VBoxHeadless
-    ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/VBoxBalloonCtrl
+    if [ -f $INSTALLATION_DIR/VBoxBalloonCtrl ]; then
+        ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/VBoxBalloonCtrl
+    fi
     if [ -f $INSTALLATION_DIR/VBoxBugReport ]; then
         ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/VBoxBugReport
     fi
